% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_mv_learners}
\alias{create_mv_learners}
\title{Make SL of multivariate learners}
\usage{
create_mv_learners(learners)
}
\arguments{
\item{learners}{List of learners supporting multivariate prediction.}
}
\description{
Make SL of multivariate learners
}
