% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data_cat_vim}
\alias{data_cat_vim}
\title{Mock data set for Variable Importance Analysis with Categorical Treatment}
\format{A \code{data.frame} with 6 columns.
\describe{
  \item{Y}{A binary variable representing an outcome of interest.}
  \item{A}{A binary variable representing an intervention of interest.}
  \item{W1}{A categorical variable representing a covariate of interest.}
  \item{W2}{A categorical variable representing a covariate of interest.}
  \item{W3}{A categorical variable representing a covariate of interest.}
  \item{W4}{A categorical variable representing a covariate of interest.}
}}
\usage{
data_cat_vim
}
\description{
A dataset with a simple data structure O = (A, Y, W), where outcome (Y) is
binary and treatment (A) is categorical. 
This is a simple dataset designed specifically to illustrate the TMLE estimation procedure.
}
\keyword{datasets}
