% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_rule.R
\docType{class}
\name{LF_rule}
\alias{LF_rule}
\title{Dynamic Likelihood Factor}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
 \code{\link{LF_base}}.
}
\value{
\code{\link[tmle3]{LF_base}} object.
}
\description{
Dynamic Likelihood Factor built on top of \code{\link[tmle3]{LF_base}}.
}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{LF_rule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LF_rule-new}{\code{LF_rule$new()}}
\item \href{#method-LF_rule-get_mean}{\code{LF_rule$get_mean()}}
\item \href{#method-LF_rule-get_density}{\code{LF_rule$get_density()}}
\item \href{#method-LF_rule-cf_values}{\code{LF_rule$cf_values()}}
\item \href{#method-LF_rule-clone}{\code{LF_rule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="delayed_train"><a href='../../tmle3/html/LF_base.html#method-LF_base-delayed_train'><code>tmle3::LF_base$delayed_train()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="get_likelihood"><a href='../../tmle3/html/LF_base.html#method-LF_base-get_likelihood'><code>tmle3::LF_base$get_likelihood()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print"><a href='../../tmle3/html/LF_base.html#method-LF_base-print'><code>tmle3::LF_base$print()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="sample"><a href='../../tmle3/html/LF_base.html#method-LF_base-sample'><code>tmle3::LF_base$sample()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="train"><a href='../../tmle3/html/LF_base.html#method-LF_base-train'><code>tmle3::LF_base$train()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LF_rule-new"></a>}}
\if{latex}{\out{\hypertarget{method-LF_rule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_rule$new(name, type = "density", rule_fun, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LF_rule-get_mean"></a>}}
\if{latex}{\out{\hypertarget{method-LF_rule-get_mean}{}}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_rule$get_mean(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LF_rule-get_density"></a>}}
\if{latex}{\out{\hypertarget{method-LF_rule-get_density}{}}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_rule$get_density(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LF_rule-cf_values"></a>}}
\if{latex}{\out{\hypertarget{method-LF_rule-cf_values}{}}}
\subsection{Method \code{cf_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_rule$cf_values(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LF_rule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LF_rule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_rule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
