% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Rule.R
\docType{data}
\name{Optimal_Rule}
\alias{Optimal_Rule}
\title{Learns the Optimal Rule given a tmle_task and likelihood}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Optimal_Rule
}
\description{
Learns the Optimal Rule given a tmle_task and likelihood
}
\keyword{datasets}
