% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_TSM_names.R
\docType{class}
\name{Param_TSM_name}
\alias{Param_TSM_name}
\title{Treatment Specific Mean with names specifying the covariates the rule depends on.}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
 \code{\link[tmle3]{Param_base}}.
}
\value{
TSP with name of the intervention specified.
}
\description{
Treatment Specific Mean with names specifying the covariates the rule depends on.
}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_TSM_name}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Param_TSM_name-new}{\code{Param_TSM_name$new()}}
\item \href{#method-Param_TSM_name-clever_covariates}{\code{Param_TSM_name$clever_covariates()}}
\item \href{#method-Param_TSM_name-estimates}{\code{Param_TSM_name$estimates()}}
\item \href{#method-Param_TSM_name-clone}{\code{Param_TSM_name$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print"><a href='../../tmle3/html/Param_base.html#method-Param_base-print'><code>tmle3::Param_base$print()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param_TSM_name-new"></a>}}
\if{latex}{\out{\hypertarget{method-Param_TSM_name-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TSM_name$new(
  observed_likelihood,
  intervention_list,
  v = NULL,
  ...,
  outcome_node = "Y"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param_TSM_name-clever_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-Param_TSM_name-clever_covariates}{}}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TSM_name$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param_TSM_name-estimates"></a>}}
\if{latex}{\out{\hypertarget{method-Param_TSM_name-estimates}{}}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TSM_name$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param_TSM_name-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Param_TSM_name-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TSM_name$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
