% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Q_learning}
\alias{Q_learning}
\title{Q learning wrapper}
\usage{
Q_learning(tmle_spec_Q, initial_likelihood, tmle_task)
}
\arguments{
\item{tmle_spec_Q}{TMLE Spec initializing Q learning.}

\item{initial_likelihood}{Initial likelihood object as defined and obtained by tmle3.}

\item{tmle_task}{TMLE task object as defined and obtained by tmle3.}
}
\description{
Q learning wrapper
}
