% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx.R
\name{tmle3_mopttx}
\alias{tmle3_mopttx}
\title{O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
V=Covariates the rule depends on}
\usage{
tmle3_mopttx(V, type, b_learner)
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
V=Covariates the rule depends on
}
