% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_blip_revere.R
\name{tmle3_mopttx_blip_revere}
\alias{tmle3_mopttx_blip_revere}
\title{Mean under the Optimal Individualized Treatment Rule}
\usage{
tmle3_mopttx_blip_revere(V = NULL, type = "blip1", learners,
  maximize = TRUE, complex = TRUE, realistic = FALSE)
}
\arguments{
\item{V}{Covariates the rule depends on.}

\item{type}{One of three psudo-blip versions developed to accommodate categorical treatment. "Blip1"
corresponds to chosing a reference category, and defining the blip for all other categories relative to the
specified reference. Note that in the case of binary treatment, "blip1" is just the usual blip.
"Blip2$ corresponds to defining the blip relative to the average of all categories. Finally,
"Blip3" corresponds to defining the blip relative to the weighted average of all categories.}

\item{learners}{Library for Y (outcome), A (treatment), and B (blip) estimation.}

\item{maximize}{Specify whether we want to maximize or minimize the mean of the final outcome.}

\item{complex}{If \code{TRUE}, learn the rule using the specified covariates \code{V}. If
\code{FALSE}, check if a less complex rule is better.}

\item{realistic}{If \code{TRUE}, it will return a rule what is possible due to practical positivity constraints.}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
