% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_Q.R
\name{tmle3_mopttx_Q}
\alias{tmle3_mopttx_Q}
\title{Mean under the Optimal Individualized Treatment Rule, using Q learning}
\usage{
tmle3_mopttx_Q(maximize)
}
\arguments{
\item{maximize}{Specify whether we want to maximize or minimize the mean of the final outcome.}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
