% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_vim.R
\name{tmle3_mopttx_vim}
\alias{tmle3_mopttx_vim}
\title{Mean under the Optimal Individualized Treatment Rule}
\usage{
tmle3_mopttx_vim(V = NULL, type = "blip2", method = "SL",
  learners = NULL, contrast = "linear", maximize = TRUE,
  complex = TRUE, realistic = FALSE)
}
\arguments{
\item{V}{Covariates the rule depends on.}

\item{type}{One of three psudo-blip versions developed to accommodate categorical treatment. "Blip1"
corresponds to chosing a reference category, and defining the blip for all other categories relative to the
specified reference. Note that in the case of binary treatment, "blip1" is just the usual blip.
"Blip2$ corresponds to defining the blip relative to the average of all categories. Finally,
"Blip3" corresponds to defining the blip relative to the weighted average of all categories.}

\item{method}{Specifies which methodology to use for learning the rule. Options are "Q" for Q-learning, and
"SL" for the Super-Learner approach using split-specific estimates.}

\item{learners}{Library for Y (outcome), A (treatment), and B (blip) estimation.}

\item{contrast}{Defined either a linear or multiplicative contrast for the delta method.}

\item{maximize}{Specify whether we want to maximize or minimize the mean of the final outcome.}

\item{complex}{If \code{TRUE}, learn the rule using the specified covariates \code{V}. If
\code{FALSE}, check if a less complex rule is better.}

\item{realistic}{If \code{TRUE}, it will return a rule what is possible due to practical positivity constraints.}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
