% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx.R
\docType{data}
\name{tmle3_Spec_mopttx}
\alias{tmle3_Spec_mopttx}
\title{Defines a tmle for the mean under the optimal individualized treatment with categorical treatment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_mopttx
}
\description{
Defines a tmle for the mean under the optimal individualized treatment with categorical treatment
}
\keyword{datasets}
