% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_blip_revere.R
\name{tmle3_Spec_mopttx_blip_revere}
\alias{tmle3_Spec_mopttx_blip_revere}
\title{Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment under Revere CV-TMLE.}
\description{
Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment under Revere CV-TMLE.

Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment under Revere CV-TMLE.
}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{tmle3_Spec_mopttx_blip_revere}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Spec_mopttx_blip_revere$new()}}
\item \href{#method-vals_from_factor}{\code{tmle3_Spec_mopttx_blip_revere$vals_from_factor()}}
\item \href{#method-predict_rule}{\code{tmle3_Spec_mopttx_blip_revere$predict_rule()}}
\item \href{#method-make_rules}{\code{tmle3_Spec_mopttx_blip_revere$make_rules()}}
\item \href{#method-make_est_fin}{\code{tmle3_Spec_mopttx_blip_revere$make_est_fin()}}
\item \href{#method-set_opt}{\code{tmle3_Spec_mopttx_blip_revere$set_opt()}}
\item \href{#method-set_rule}{\code{tmle3_Spec_mopttx_blip_revere$set_rule()}}
\item \href{#method-data_adapt_psi}{\code{tmle3_Spec_mopttx_blip_revere$data_adapt_psi()}}
\item \href{#method-make_params}{\code{tmle3_Spec_mopttx_blip_revere$make_params()}}
\item \href{#method-clone}{\code{tmle3_Spec_mopttx_blip_revere$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_initial_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_initial_likelihood}{\code{tmle3::tmle3_Spec$make_initial_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_targeted_likelihood}{\code{tmle3::tmle3_Spec$make_targeted_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_tmle_task">}\href{../../tmle3/html/tmle3_Spec.html#method-make_tmle_task}{\code{tmle3::tmle3_Spec$make_tmle_task()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_updater">}\href{../../tmle3/html/tmle3_Spec.html#method-make_updater}{\code{tmle3::tmle3_Spec$make_updater()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$new(
  V = NULL,
  type,
  learners,
  maximize = TRUE,
  complex = TRUE,
  realistic = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vals_from_factor"></a>}}
\if{latex}{\out{\hypertarget{method-vals_from_factor}{}}}
\subsection{Method \code{vals_from_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$vals_from_factor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_rule"></a>}}
\if{latex}{\out{\hypertarget{method-predict_rule}{}}}
\subsection{Method \code{predict_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$predict_rule(tmle_task_new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_rules"></a>}}
\if{latex}{\out{\hypertarget{method-make_rules}{}}}
\subsection{Method \code{make_rules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_rules(V)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_est_fin"></a>}}
\if{latex}{\out{\hypertarget{method-make_est_fin}{}}}
\subsection{Method \code{make_est_fin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_est_fin(fit, max, p.value = 0.35)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opt"></a>}}
\if{latex}{\out{\hypertarget{method-set_opt}{}}}
\subsection{Method \code{set_opt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$set_opt(opt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rule"></a>}}
\if{latex}{\out{\hypertarget{method-set_rule}{}}}
\subsection{Method \code{set_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$set_rule(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_adapt_psi"></a>}}
\if{latex}{\out{\hypertarget{method-data_adapt_psi}{}}}
\subsection{Method \code{data_adapt_psi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$data_adapt_psi(data_tda, node_list, Qbar0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_params"></a>}}
\if{latex}{\out{\hypertarget{method-make_params}{}}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_params(tmle_task, likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
