% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_rule.R
\docType{data}
\name{LF_rule}
\alias{LF_rule}
\title{Dynamic Likelihood Factor}
\format{\code{\link{R6Class}} object.}
\usage{
LF_rule
}
\value{
\code{LF_base} object
}
\description{
Dynamic Likelihood Factor
}
\section{Constructor}{

  \code{define_lf(LF_static, name, type, value, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name
    in the nodes specified by tmle3_Task.}

    \item{\code{type}}{character, either 'density', for conditional density or, 'mean' for conditional mean
    }
    \item{\code{value}}{the static value
    }
    \item{\code{...}}{Not currently used.
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{value}}{the static value.}
    }
}

\concept{Likelihood objects}
\keyword{data}
