% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Rule_Q_learning.R
\docType{data}
\name{Optimal_Rule_Q_learning}
\alias{Optimal_Rule_Q_learning}
\title{Learns the Optimal Rule given a tmle_task and likelihood, using Q learning.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Optimal_Rule_Q_learning
}
\description{
Learns the Optimal Rule given a tmle_task and likelihood, using Q learning.
}
\keyword{datasets}
