% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Rule_Revere.R
\docType{data}
\name{Optimal_Rule_Revere}
\alias{Optimal_Rule_Revere}
\title{Learns the Optimal Rule given a tmle_task and likelihood}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Optimal_Rule_Revere
}
\description{
Uses new revere framework
}
\keyword{datasets}
