% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Rule_Revere.R
\docType{class}
\name{Optimal_Rule_Revere}
\alias{Optimal_Rule_Revere}
\title{Learning the Optimal Rule using the Revere framework}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
 \code{\link{tmle3_Spec}}.
}
\value{
A optimal rule object inheriting from \code{\link{tmle3_Spec}} with
methods for learning the optimal rule. For a full list of the available
functionality, see the complete documentation of \code{\link{tmle3_Spec}}.
}
\description{
Functions used to learn the Optimal Rule given a tmle_task and likelihood,
using the Revere framework. Complements 'tmle3_Spec_mopttx_blip_revere' class.
}
\section{Parameters}{

  - \code{tmle_task}: Task object specifying the data and node structure.
  - \code{tmle_spec}: Spec object of \code{\link[keras]{tmle3}}. Allows for different
  Specs to use the current class for learning the Optimal Rule.
  - \code{likelihood}: Likelihood object of \code{\link[keras]{tmle3}}, corresponding
  to the current estimate of the required parts of the likelihood necessary for the target
  parameter.
  - \code{V}: User-specified list of covariates used to define the rule.
  - \code{blip_type}: Blip type, corresponding to different ways of defining the
  reference category in learning the blip; mostly applies to categorical treatment.
  Available categories include "blip1" (reference level of treatment), "blip2"
  (average level of treatment) and "blip3" (weighted average level of treatment).
  - \code{learners}: List of user-defined learners for relevant parts of the
  likelihood.
  - \code{maximize}: Should the average outcome be maximized of minimized? Default is
  maximize=TRUE.
  - \code{shift_grid}: Grid of possible values for the stochastic optimal rule.
  Work in progress.
}

\examples{
\dontrun{
library(sl3)
library(tmle3)
library(data.table)

data("data_bin")
data <- data_bin

Q_lib <- make_learner_stack("Lrnr_mean", "Lrnr_glm_fast")
g_lib <- make_learner_stack("Lrnr_mean", "Lrnr_glm_fast")
B_lib <- make_learner_stack("Lrnr_glm_fast", "Lrnr_xgboost")

metalearner <- make_learner(Lrnr_nnls)
Q_learner <- make_learner(Lrnr_sl, Q_lib, metalearner)
g_learner <- make_learner(Lrnr_sl, g_lib, metalearner)
B_learner <- make_learner(Lrnr_sl, B_lib, metalearner)

learner_list <- list(Y = Q_learner, A = g_learner, B = B_learner)

node_list <- list(W = c("W1", "W2", "W3"), A = "A", Y = "Y")

tmle_spec <- tmle3_mopttx_blip_revere(
  V = c("W1", "W2", "W3"),
  type = "blip1", learners = learner_list, maximize = TRUE,
  complex = TRUE, realistic = TRUE
)
}
}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{Optimal_Rule_Revere}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Optimal_Rule_Revere$new()}}
\item \href{#method-factor_to_indicators}{\code{Optimal_Rule_Revere$factor_to_indicators()}}
\item \href{#method-V_data}{\code{Optimal_Rule_Revere$V_data()}}
\item \href{#method-DR_full}{\code{Optimal_Rule_Revere$DR_full()}}
\item \href{#method-blip_revere_function}{\code{Optimal_Rule_Revere$blip_revere_function()}}
\item \href{#method-bound}{\code{Optimal_Rule_Revere$bound()}}
\item \href{#method-fit_blip}{\code{Optimal_Rule_Revere$fit_blip()}}
\item \href{#method-rule}{\code{Optimal_Rule_Revere$rule()}}
\item \href{#method-rule_stochastic}{\code{Optimal_Rule_Revere$rule_stochastic()}}
\item \href{#method-clone}{\code{Optimal_Rule_Revere$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_initial_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_initial_likelihood}{\code{tmle3::tmle3_Spec$make_initial_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_params">}\href{../../tmle3/html/tmle3_Spec.html#method-make_params}{\code{tmle3::tmle3_Spec$make_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_targeted_likelihood}{\code{tmle3::tmle3_Spec$make_targeted_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_tmle_task">}\href{../../tmle3/html/tmle3_Spec.html#method-make_tmle_task}{\code{tmle3::tmle3_Spec$make_tmle_task()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_updater">}\href{../../tmle3/html/tmle3_Spec.html#method-make_updater}{\code{tmle3::tmle3_Spec$make_updater()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$new(
  tmle_task,
  tmle_spec,
  likelihood,
  V = NULL,
  blip_type = "blip2",
  learners,
  maximize = TRUE,
  shift_grid = seq(-1, 1, by = 0.5)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-factor_to_indicators"></a>}}
\if{latex}{\out{\hypertarget{method-factor_to_indicators}{}}}
\subsection{Method \code{factor_to_indicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$factor_to_indicators(x, x_vals)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V_data"></a>}}
\if{latex}{\out{\hypertarget{method-V_data}{}}}
\subsection{Method \code{V_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$V_data(tmle_task, fold = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DR_full"></a>}}
\if{latex}{\out{\hypertarget{method-DR_full}{}}}
\subsection{Method \code{DR_full()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$DR_full(v, indx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blip_revere_function"></a>}}
\if{latex}{\out{\hypertarget{method-blip_revere_function}{}}}
\subsection{Method \code{blip_revere_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$blip_revere_function(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bound"></a>}}
\if{latex}{\out{\hypertarget{method-bound}{}}}
\subsection{Method \code{bound()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$bound(cv_g)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_blip"></a>}}
\if{latex}{\out{\hypertarget{method-fit_blip}{}}}
\subsection{Method \code{fit_blip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$fit_blip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rule"></a>}}
\if{latex}{\out{\hypertarget{method-rule}{}}}
\subsection{Method \code{rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$rule(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rule_stochastic"></a>}}
\if{latex}{\out{\hypertarget{method-rule_stochastic}{}}}
\subsection{Method \code{rule_stochastic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$rule_stochastic(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimal_Rule_Revere$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
