% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Rule_Revere.R
\docType{data}
\name{Optimal_Rule_Revere}
\alias{Optimal_Rule_Revere}
\title{Learns the Optimal Rule given a tmle_task and likelihood, using the Revere framework.
Complements 'tmle3_Spec_mopttx_blip_revere'.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Optimal_Rule_Revere
}
\description{
Learns the Optimal Rule given a tmle_task and likelihood, using the Revere framework.
Complements 'tmle3_Spec_mopttx_blip_revere'.
}
\keyword{datasets}
