% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Q_learning}
\alias{Q_learning}
\title{Q learning wrapper}
\usage{
Q_learning(tmle_spec_Q, learner_list, B = 1000, data, node_list)
}
\arguments{
\item{tmle_spec_Q}{TMLE Spec initializing Q learning.}

\item{learner_list}{List of algorithms used to fit Q}

\item{B}{Number of bootstraps}

\item{data}{Dataset used}

\item{node_list}{List of nodes corresponding to Y, A and W.}
}
\description{
Q learning wrapper
}
