% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_blip_revere.R
\docType{class}
\name{tmle3_Spec_mopttx_blip_revere}
\alias{tmle3_Spec_mopttx_blip_revere}
\title{TMLE for the Mean Under the Optimal Individualized Rule}
\format{
An \code{\link[R6]{R6Class}} object inheriting from
 \code{\link{tmle3_Spec}}.
}
\value{
A tmle3 object inheriting from \code{\link{tmle3_Spec}} with
methods for obtaining the TMLE for the Mean Under the Optimal Individualized Rule.
For a full list of the available  functionality, see the complete documentation of
\code{\link{tmle3_Spec}}.
}
\description{
The functions contained in the class define a TMLE for the Mean Under
the Optimal Individualized Rule with Categorical Treatment, learned and estimated
under Revere CV-TMLE. For learning the Optimal Rule, see 'Optimal_Rule_Revere' class.
}
\section{Parameters}{

  - \code{V}: User-specified list of covariates used to define the rule.
  - \code{type}: Blip type, corresponding to different ways of defining the
  reference category in learning the blip; mostly applies to categorical treatment.
  Available categories include "blip1" (reference level of treatment), "blip2"
  (average level of treatment) and "blip3" (weighted average level of treatment).
  - \code{learners}: List of user-defined learners for relevant parts of the
  likelihood.
  - \code{maximize}: Should the average outcome be maximized of minimized? Default is
  maximize=TRUE.
  - \code{complex}: If \code{TRUE}, the returned mean under the Optimal Rule is based on the
  full set of covariates provided by the user (parameter "V"). If \code{FALSE}, simpler rules
  (including the static rules), are evaluated as well; the returned mean under the Optimal
  Rule is then a potentially more parsimonious rule, if the mean performance is similar.
  - \code{realistic}: If \code{TRUE}, the optimal rule returned takes into account the
  probability of treatment given covariates.
  - \code{resource}: Indicates the percent of initially estimated individuals who should be given 
  treatment that get treatment, based on their blip estimate. If resource = 1 all estimated 
  individuals to benefit from treatment get treatment, if resource = 0 none get treatment.
}

\examples{
library(sl3)
library(tmle3)
library(data.table)

data("data_bin")
data <- data_bin

Q_lib <- make_learner_stack("Lrnr_mean", "Lrnr_glm_fast")
g_lib <- make_learner_stack("Lrnr_mean", "Lrnr_glm_fast")
B_lib <- make_learner_stack("Lrnr_glm_fast", "Lrnr_xgboost")

metalearner <- make_learner(Lrnr_nnls)
Q_learner <- make_learner(Lrnr_sl, Q_lib, metalearner)
g_learner <- make_learner(Lrnr_sl, g_lib, metalearner)
B_learner <- make_learner(Lrnr_sl, B_lib, metalearner)

learner_list <- list(Y = Q_learner, A = g_learner, B = B_learner)

node_list <- list(W = c("W1", "W2", "W3"), A = "A", Y = "Y")

tmle_spec <- tmle3_mopttx_blip_revere(
  V = c("W1", "W2", "W3"),
  type = "blip1", learners = learner_list, maximize = TRUE,
  complex = TRUE, realistic = TRUE
)

fit <- tmle3(tmle_spec, data, node_list, learner_list)
fit$summary
}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{tmle3_Spec_mopttx_blip_revere}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Spec_mopttx_blip_revere$new()}}
\item \href{#method-vals_from_factor}{\code{tmle3_Spec_mopttx_blip_revere$vals_from_factor()}}
\item \href{#method-predict_rule}{\code{tmle3_Spec_mopttx_blip_revere$predict_rule()}}
\item \href{#method-make_rules}{\code{tmle3_Spec_mopttx_blip_revere$make_rules()}}
\item \href{#method-make_est_fin}{\code{tmle3_Spec_mopttx_blip_revere$make_est_fin()}}
\item \href{#method-set_opt}{\code{tmle3_Spec_mopttx_blip_revere$set_opt()}}
\item \href{#method-set_rule}{\code{tmle3_Spec_mopttx_blip_revere$set_rule()}}
\item \href{#method-data_adapt_psi}{\code{tmle3_Spec_mopttx_blip_revere$data_adapt_psi()}}
\item \href{#method-make_params}{\code{tmle3_Spec_mopttx_blip_revere$make_params()}}
\item \href{#method-clone}{\code{tmle3_Spec_mopttx_blip_revere$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_initial_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_initial_likelihood}{\code{tmle3::tmle3_Spec$make_initial_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_targeted_likelihood}{\code{tmle3::tmle3_Spec$make_targeted_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_tmle_task">}\href{../../tmle3/html/tmle3_Spec.html#method-make_tmle_task}{\code{tmle3::tmle3_Spec$make_tmle_task()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_updater">}\href{../../tmle3/html/tmle3_Spec.html#method-make_updater}{\code{tmle3::tmle3_Spec$make_updater()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$new(
  V = NULL,
  type,
  learners,
  maximize = TRUE,
  complex = TRUE,
  realistic = FALSE,
  resource = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vals_from_factor"></a>}}
\if{latex}{\out{\hypertarget{method-vals_from_factor}{}}}
\subsection{Method \code{vals_from_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$vals_from_factor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_rule"></a>}}
\if{latex}{\out{\hypertarget{method-predict_rule}{}}}
\subsection{Method \code{predict_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$predict_rule(tmle_task_new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_rules"></a>}}
\if{latex}{\out{\hypertarget{method-make_rules}{}}}
\subsection{Method \code{make_rules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_rules(V)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_est_fin"></a>}}
\if{latex}{\out{\hypertarget{method-make_est_fin}{}}}
\subsection{Method \code{make_est_fin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_est_fin(fit, max, p.value = 0.35)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opt"></a>}}
\if{latex}{\out{\hypertarget{method-set_opt}{}}}
\subsection{Method \code{set_opt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$set_opt(opt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rule"></a>}}
\if{latex}{\out{\hypertarget{method-set_rule}{}}}
\subsection{Method \code{set_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$set_rule(rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_adapt_psi"></a>}}
\if{latex}{\out{\hypertarget{method-data_adapt_psi}{}}}
\subsection{Method \code{data_adapt_psi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$data_adapt_psi(data_tda, node_list, Qbar0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_params"></a>}}
\if{latex}{\out{\hypertarget{method-make_params}{}}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$make_params(tmle_task, likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_mopttx_blip_revere$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
