% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_blip_revere.R
\docType{data}
\name{tmle3_Spec_mopttx_blip_revere}
\alias{tmle3_Spec_mopttx_blip_revere}
\title{Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment under Revere CV-TMLE.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
tmle3_Spec_mopttx_blip_revere
}
\description{
Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment under Revere CV-TMLE.
}
\keyword{datasets}
