% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_Q.R
\docType{data}
\name{tmle3_Spec_mopttx_Q}
\alias{tmle3_Spec_mopttx_Q}
\title{Defines the Mean Under the Optimal Individualized Rule with Categorical Treatment,
estimated using Q learning (single step).}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
tmle3_Spec_mopttx_Q
}
\description{
Defines the Mean Under the Optimal Individualized Rule with Categorical Treatment,
estimated using Q learning (single step).
}
\keyword{datasets}
