% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_mopttx_vim.R
\docType{data}
\name{tmle3_Spec_mopttx_vim}
\alias{tmle3_Spec_mopttx_vim}
\title{Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment and contrast with the Mean under Observed Treatment.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
tmle3_Spec_mopttx_vim
}
\description{
Defines a TMLE for the Mean Under the Optimal Individualized Rule with
Categorical Treatment and contrast with the Mean under Observed Treatment.
}
\keyword{datasets}
